requirements.txt
mysqlclient==2.1.0
Django==3.2.12
requests==2.27.1
beautifulsoup4==4.10.0
django-background-tasks==1.2.5
pytz==2021.3
google-api-python-client==2.37.0
